REST APIs: Adding Authorization Checks

+++ rest-project/backend/controllers/feed.js
@@ -134,6 +134,12 @@ exports.updatePost = (req, res, next) => {
         throw error;
       }
 
+      if(post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!');
+        error.statusCode = 403;
+        throw error;
+      }
+
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -166,6 +172,13 @@ exports.deletePost = (req, res, next) => {
         err.statusCode = 422;
         throw error;
       }
+
+      if(post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!');
+        error.statusCode = 403;
+        throw error;
+      }
+
       clearImage(post.imageUrl);
       return post.delete();
     })
